pyproject.toml
[project]
name = "pointcalculatorbackend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.128.4",
    "gotrue>=2.12.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "supabase>=2.27.3",
    "uvicorn>=0.40.0",
]
